Connected /newevent frontend to /evenlist backend, reorganized some files
