+++ backend/dummy/events.json
@@ -1,22 +1,48 @@
-{
-    "events": [
-        {
-            "name": "Tech Conference 2024",
-            "date": "2024-10-15",
-            "location": "San Francisco, CA",
-            "description": "Join us for a day of insightful talks and networking with industry leaders in technology."
-        },
-        {
-            "name": "Music Festival",
-            "date": "2024-09-25",
-            "location": "Austin, TX",
-            "description": "A weekend filled with live music performances from top artists around the world."
-        },
-        {
-            "name": "Startup Pitch Night",
-            "date": "2024-11-05",
-            "location": "New York, NY",
-            "description": "Watch innovative startups pitch their ideas to investors and compete for prizes."
-        }
-    ]
-}
+[
+    {
+        "name": "Tech Conference 2024",
+        "date": "2024-10-15",
+        "location": "San Francisco, CA",
+        "description": "Join us for a day of insightful talks and networking with industry leaders in technology."
+    },
+    {
+        "name": "Music Festival",
+        "date": "2024-09-25",
+        "location": "Austin, TX",
+        "description": "A weekend filled with live music performances from top artists around the world."
+    },
+    {
+        "name": "Startup Pitch Night",
+        "date": "2024-11-05",
+        "location": "New York, NY",
+        "description": "Watch innovative startups pitch their ideas to investors and compete for prizes."
+    },
+    {
+        "id": 1,
+        "name": "Event",
+        "description": "This is a test event.",
+        "location": "New York",
+        "skills": "Python, Flask",
+        "urgency": "HIGH",
+        "date": "2024-10-05T10:00:00"
+    },
+    {
+        "id": 1,
+        "name": "Event",
+        "description": "This is a test event.",
+        "location": "New York",
+        "skills": "Python, Flask",
+        "urgency": "HIGH",
+        "date": "2024-10-05T10:00:00"
+    },
+    {
+        "name": "dqd",
+        "description": "dqwdqw",
+        "location": "dqwwq",
+        "skills": [
+            "goodwithchildren"
+        ],
+        "urgency": "1",
+        "date": "2024-10-07T16:03:53"
+    }
+]
