--- frontend/src/routes/EventPage.jsx
@@ -1,16 +1,0 @@
-function EventPage() {
-  return (
-    <>
-      <header><h1>Events</h1></header>
-      <h2>List of Events by Area</h2>
-      <p><a href="">Cypress</a></p>
-      <p><a href="">Houston downtown area</a></p>
-      <p><a href="">Katy</a></p>
-      <p><a href="">League City</a></p>
-      <p><a href="">Pearland</a></p>
-      <p><a href="">Sugarland</a></p>
-      <p><a href="">Woodland</a></p>
-    </>
-  )
-}
-export default EventPage
